overrides: prefer env attr
Nixpkgs is moving towards preferring this. In general, it is clearer as
it more clearly separates attrs that are solely for the build env vs.
intending to control the builder or mkDerivation. Blanket changed
anything matching `[A-Z][A-Z0-9_]+ =`.

This can become a problem (in either direction) in some cases if the
override here is overlaid atop a nixpkgs derivation that defines
differently, with e.g.

> error: The ‘env’ attribute set cannot contain any attributes passed to
derivation. The following attributes are overlapping: …